chore: refactor getFormSchema
