+++ src/app/form/done/page.tsx
@@ -4,10 +4,10 @@ export default function Page() {
   return (
     <div className='flex w-full flex-col gap-5 rounded-box border-2 p-10'>
       <h1 className='text-4xl font-bold text-neutral-900'>
-        คุณส่งไฟล์เรียบร้อย
+        คำตอบของคุณถูกบันทึกไว้แล้ว
       </h1>
       <hr />
-      <h2 className='text-lg text-neutral-500'>ขอบคุณที่ส่ง</h2>
+      <h2 className='text-lg text-neutral-500'>ขอบคุณที่ส่งฟอร์ม</h2>
     </div>
   )
 }
